Fix workload unsetting bug
When navigating through the modal, it was possible to accidentally unset
your workload by either skipping the workload page, or clicking "next"
very quickly through it.

This changes that state to be managed at the modal level, instead of on
the specific workload page (which may not ever render)
